requirements.txt
Django==3.2.4
django-extensions==3.1.3
djangorestframework==3.12.4
drf_yasg==1.20.0
# Found djangorestframework-dataclasses in https://github.com/encode/django-rest-framework/issues/6021
djangorestframework-dataclasses==0.9
# git+ssh://git@github.com/oxan/djangorestframework-dataclasses.git#egg=djangorestframework-dataclasses
djangorestframework-camel-case==1.2.0
pre-commit==2.13.0
